M3 Source code
The zip file contains the source code, executable and sample of input and output files for M3. 

M3 can be run as follows:
- If M3 is used for the first time, unpack M3_source_code.zip at a chosen location.
- Navigate to ...\M3_source_code\M3\RunSimulationsFromInputFile\bin\Debug
- Run the file "RunSimulationsFromInputFile.exe" 

This will run all the simulations that are specified in the file "...\M3_source_code\M3\RunSimulationsFromInputFile\bin\Debug\simulationInput.csv". The output files from the simulations will be saved in the folder  ""...\M3_source_code\M3\RunSimulationsFromInputFile\bin\Debug\Output files".
M3 source.zip contains, besides the model, also a number of input files for sample simulations. More information on the structure of these input files can be found in Section 8 in the Technical description.